Renaming to conform to M1000-0001-091
M1000-0001-091 describes terminology, and this brings katcbfsim into
line with it. This means:
- renaming "stream" to "substream" when it refers to the per-engine data
- renaming "product" to "stream"
- renaming "stream" to "transport" ("transport" isn't defined in the
  document, but it needed a new name to disambiguate it, and "transport"
  is a more descriptive name)

There are also some bonus commits:
- katcp 0.6 is released, so this is now specified as a minimum
  requirement in setup.py, and docs updated
- ?gain and ?n-dumps commands should now actually work
- a few PEP8 indentation fixes
- commands.py has been abused to improve coverage

This is a backwards-compatible change due to the name changes of the
command-line arguments and katcp commands. Do not merge without the
corresponding pull request on the master controller (TBD).
