+++ setup.py
@@ -13,7 +13,7 @@ setup(
     description='Simulator for MeerKAT correlator',
     scripts=['scripts/cbfsim.py'],
     setup_requires=['katversion'],
-    install_requires=['katcp', 'trollius', 'tornado>=4.2', 'spead2>=1.1.0', 'katpoint', 'numpy', 'h5py',
+    install_requires=['katcp>=0.6.0', 'trollius', 'tornado>=4.2', 'spead2>=1.1.0', 'katpoint', 'numpy', 'h5py',
         'katsdpsigproc[CUDA]', 'katsdptelstate'],
     tests_require=tests_require,
     extras_require={'test': tests_require, 'doc': ['sphinx>=1.3']},
